Finished game pull function and First transformation
The Function to pull the data From one Game and transform it into a coherent DataFrame with the proper type conversions has been finished. Also along with it some Helper Functions to extract certain
data which is subject to be expanded and improved, as right now it is a bit ineffcient still.
Also The calculation for average player stats has been added. Here the shift down one for games still needs to be implemented.